fix(badge): demo拆解与规范
